添加其他的测试文件

+++ src/cacheline_test.c
@@ -1,4 +1,3 @@
-
 #include <stdio.h>
 #include <stdlib.h>
 #include <time.h>
@@ -18,7 +17,7 @@ int main() {
         }
         end = clock();
         double time_ms = 1000.0 * (end - start) / CLOCKS_PER_SEC;
-        printf("%d\t%.3f\n", stride * sizeof(int), time_ms);
+        printf("%lu\t%.3f\n", (unsigned long)(stride * sizeof(int)), time_ms);
     }
 
     return sum; // use sum to prevent optimization
